// wgg_front/src/lib/api/graphql_types/graphql.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * Implement the DateTime<Utc> scalar
   *
   * The input/output is a string in RFC3339 format.
   */
  DateTime: any;
};
// ...
export type AggregateProductInput = {
  aggregateId: Scalars['Int'];
  quantity: Scalars['Int'];
};
// ...
export type NoteProductInput = {
  content: Scalars['String'];
  quantity: Scalars['Int'];
};
// ...
export enum Provider {
  Jumbo = 'JUMBO',
  Picnic = 'PICNIC'
}
// ...
export type RawProductInput = {
  productId: Scalars['String'];
  provider: Provider;
  quantity: Scalars['Int'];
};